edit admin item index bar
