+++ app/views/admin/items/index.html.erb
@@ -7,7 +7,7 @@
   </div>
 
   <table class="table">
-    <tr><th>Item</th><th></th><th>Description</th><th>Price</th><th></th></tr>
+    <tr><th>Status</th> <th>Item</th><th></th><th>Description</th><th>Price</th><th></th></tr>
     <% @items.each do |item| %>
       <tr>
         <td>
