+++ spec/features/admin_items_spec.rb
@@ -16,11 +16,10 @@ feature "Admin User Items" do
   end
 
   scenario "User Admin can visit items index page" do
-    Item.create(title: "food", description: "Yummy")
     allow_any_instance_of(ApplicationController).
       to receive(:current_user).and_return(@admin_user)
-
     visit admin_items_path
+    expect(page).to have_content("Hot n Spicy")
   end
 
   scenario "Admin User can create new items and see them on index page" do
@@ -53,6 +52,22 @@ feature "Admin User Items" do
     expect(page).to have_content("Delish")
   end
 
+  scenario "Admin User can edit current items and see them updated on index page" do
+    allow_any_instance_of(ApplicationController).to receive(:current_user).and_return(@admin_user)
+    visit admin_items_path
+    click_link_or_button "Edit Item"
+    expect(current_path).to eq(edit_admin_item_path(@item))
+    fill_in "item[title]", with: "new"
+    fill_in "item[description]", with: "this is it"
+    fill_in "item[price]", with: "2200"
+    find(:css, "#category_list_categories_meat[value='Meat']").set(true)
+    click_link_or_button "Save"
+    expect(current_path).to eq(admin_items_path)
+    expect(page).to have_content("new")
+    expect(page).to have_content("this is it")
+    expect(page).to have_content("$22.00")
+  end
+
   scenario "Admin User can edit an item and reassign it to a category" do
     allow_any_instance_of(ApplicationController).
       to receive(:current_user).and_return(@admin_user)
@@ -61,6 +76,23 @@ feature "Admin User Items" do
     fill_in "item[title]", with: "Hot n Spicy"
     fill_in "item[description]", with: "It's the best"
     fill_in "item[price]", with: "1234"
+    find(:css, "#category_list_categories_meat[value='Meat']").set(true)
+    click_link_or_button "Save"
+    expect(current_path).to eq(admin_items_path)
+    click_link "Meat"
+    expect(page).to have_content("It's the best")
+  end
+
+  scenario "Admin User can edit an item and change its status to retired check on views" do
+    allow_any_instance_of(ApplicationController).
+    to receive(:current_user).and_return(@admin_user)
+    visit admin_items_path
+    click_link_or_button "Edit Item"
+    fill_in "item[title]", with: "Hot n Spicy"
+    fill_in "item[description]", with: "It's the best"
+    fill_in "item[price]", with: "1234"
+    find(:css, "#category_list_categories_meat[value='Meat']").set(true)
+    save_and_open_page
     find(:css, "#category_list_categories_meat[value='Meat']").set(true)
     click_link_or_button "Save"
     expect(current_path).to eq(admin_items_path)
